requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
sentence-transformers
pandas
numpy
requests
ollama
